+++ server/saved_objects/index.ts
@@ -4,3 +4,4 @@
  */
 
 export { mapSavedObjectsType } from './map_saved_object';
+export { mapIconSavedObjectsType, MAP_ICON_SAVED_OBJECT_TYPE } from './map_icon_saved_object';